requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
email-validator
psycopg2

# Local dev (SQLite - no extra driver needed, built into Python)

# Production (uncomment when deploying to Supabase/Render)
# psycopg2-binary